+++ src/main/java/net/sf/jabref/gui/keyboard/KeyBindingsDialogView.java
@@ -17,7 +17,7 @@ public class KeyBindingsDialogView extends FXMLView {
     }
 
     public void show(KeyBindingPreferences keyBindingPreferences) {
-        FXAlert keyBindingsDialog = new FXAlert(AlertType.INFORMATION, Localization.lang("Key bindings"));
+        FXAlert keyBindingsDialog = new FXAlert(AlertType.INFORMATION, Localization.lang("Key bindings"),true);
         keyBindingsDialog.setDialogPane((DialogPane) this.getView());
         KeyBindingsDialogViewModel controller = (KeyBindingsDialogViewModel) fxmlLoader.getController();
         controller.setKeyBindingPreferences(keyBindingPreferences);