+++ src/main/java/net/sf/jabref/logic/error/MessagePriority.java
@@ -1,0 +1,5 @@
+package net.sf.jabref.logic.error;
+
+public enum MessagePriority {
+    LOW, MEDIUM, HIGH;
+}