+++ src/main/java/net/sf/jabref/logic/error/TeeStream.java
@@ -18,6 +18,8 @@ package net.sf.jabref.logic.error;
 
 import java.io.PrintStream;
 
+import net.sf.jabref.logic.logging.ObservableMessages;
+
 /**
  * All writes to this print stream are copied to two print streams
  * <p/>
@@ -26,10 +28,12 @@ import java.io.PrintStream;
 public class TeeStream extends PrintStream {
 
     private final PrintStream outStream;
+    private final MessagePriority priority;
 
-    public TeeStream(PrintStream out1, PrintStream out2) {
+    public TeeStream(PrintStream out1, PrintStream out2, MessagePriority priority) {
         super(out1);
         this.outStream = out2;
+        this.priority = priority;
     }
 
     @Override
@@ -37,6 +41,11 @@ public class TeeStream extends PrintStream {
         try {
             super.write(buf, off, len);
             outStream.write(buf, off, len);
+            String s = new String(buf, off, len);
+            if (!s.equals(System.lineSeparator())) {
+                ObservableMessageWithPriority messageWithPriority = new ObservableMessageWithPriority(s.replaceAll(System.lineSeparator(), ""), priority);
+                ObservableMessages.INSTANCE.add(messageWithPriority);
+            }
         } catch (Exception ignored) {
             // Ignored
         }